// src/types/WfsMelding.ts
export type WfsMelding = {
    meldnr: string;
    bps: string;
    wegnr: string;
    hectomtr: string;
    hectoltr: string;
    incident_type: string;
    berger: string;
    rayon: string;
    melder: string;
    datum: string;
    tijdstip: string;
    latitude: number;
    longitude: number;
    gid: number;
    aankomst: string;
    nummer: string;
};